requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1
python-dotenv==1.0.0
requests==2.31.0

# Data processing
pandas==2.1.3
numpy==1.24.3

# GitHub and Slack APIs
PyGithub==2.1.1
slack-sdk==3.23.0

# Vector database and embeddings
chromadb==0.4.18
openai==1.3.5

# Text processing
tiktoken==0.5.1
nltk==3.8.1

# Utilities
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.0
